Refactor: Remove calendar router, schema, and tool implementations
