+++ backend/app/tools/calendar/__init__.py
@@ -1,0 +1,13 @@
+from .service import (
+    build_calendar_intro,
+    build_initial_calendar_metadata,
+    run_calendar_pipeline,
+)
+from .router import detect_calendar_intent_via_llm
+
+__all__ = [
+    "build_calendar_intro",
+    "build_initial_calendar_metadata",
+    "detect_calendar_intent_via_llm",
+    "run_calendar_pipeline",
+]
